// src/interface/interface.ts
export interface IProps {
  id?: number;
  date?: string;
  excerpt?: any;
  protected?: boolean;
  categories?: number;
  title?: any;
  _embedded?: any;
  content?: any;
  yoast_head_json?: any;
  replies?: [];
  slug?: React.ReactNode;
  rendered?: any;
  modified?: string;
  author_name?: [];
  author_avatar_urls?: any;
}